% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental.R
\docType{data}
\name{treepositions}
\alias{treepositions}
\title{Tree Positions}
\format{
A data frame with 151 rows and 6 variables:
\describe{
\item{id_full}{unique ID of the tree}
\item{tree}{tree species}
\item{id}{ID number of tree}
\item{style}{CSS position styling information}
\item{top}{number for percent from top}
\item{left}{number for percent from left}
}
}
\source{
\url{https://www.bbg.org/collections/cherries}
}
\usage{
treepositions
}
\description{
Coordinates to help visualize where each tree is in the gardens.
}
\examples{
treepositions
}
\keyword{datasets}
