% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\docType{data}
\name{bbgdata}
\alias{bbgdata}
\title{Brooklyn Botanic Gardens Cherry Blossom Status}
\format{
A data frame with 27 rows and 3 variables:
\describe{
\item{date}{date of bloom status}
\item{alt}{full genus and species tree name}
\item{tree}{tree species}
\item{id}{unique integer ID}
\item{bloom}{one of four bloom states: "Prebloom", "First Bloom",
"Peak Bloom", "Post-Peak Bloom"}
}
}
\source{
\url{https://www.bbg.org/collections/cherries}
}
\usage{
bbgdata
}
\description{
A data set containing historical data for the Brooklyn Botanic Garden.
}
\details{
The data serves as an archive for the more visual map of the cherry blossom
trees that the official website shows.
}
\examples{
bbgdata
}
\keyword{datasets}
