% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot garden map with bloom statuses}
\usage{
plot_map(date)
}
\arguments{
\item{date}{Date to plot. Defaults to the most recent date in the data.}
}
\value{
A ggplot object.
}
\description{
Recreate the garden map with bloom statuses for a given date. If no date is
provided, defaults to the most recent date in the data.
}
\examples{
plot_map()
plot_map("2025-04-14")  # Plot a specific date with data
plot_map("2025-01-01")  # This will throw an error since the date is not in the data
}
